requirements.txt
asgiref==3.6.0
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
